Refactor the config classes and move into a separate package
* Clean up logging and code and improve typing where possible
* Define the name for the static closure as a public variable, this is now 'luceneIndexing'
this makes removes any issues implementing code has where search() is confused with the search closure.
It is also slightly better definition for the closure
* Only add `indexed()` to non-abstract classes, this prevents hibernatesearch from throwing out unnecessary warnings
abstract classes can still be used to define luceneIndexing and also searched on.
* Due to the way the hibernate search API works, there is no need to keep setting the mapping after each call
therefore define everything contextually from the entitymapping and just keep adding to it
* Add normalizer option
* Add a sortable entry to the entity config which either takes a map or a string/boolean
If not a map then just add sortableField on the current field, otherwise define a new field for
the same property and add that as the sortableField. Due to tokenisation, sortable fields should be assigned
a normalizer rather than an analyzer but hibernatesearch will output warnings accordingly